# openapi-full.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.4"
  },
  "host": "bggiguzhkdxfwgfbuqfy.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/loans": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.loans.id"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.userId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.groupId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.amount"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.reason"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.duration"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.interestRate"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.status"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.createdAt"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/loans"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "loans"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.loans"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "loans"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.loans.id"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.userId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.groupId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.amount"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.reason"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.duration"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.interestRate"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.status"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.createdAt"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "loans"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.loans.id"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.userId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.groupId"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.amount"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.reason"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.duration"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.interestRate"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.status"
          },
          {
            "$ref": "#/parameters/rowFilter.loans.createdAt"
          },
          {
            "$ref": "#/parameters/body.loans"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "loans"
        ]
      }
    },
...
  },
  "definitions": {
...
    "loans": {
      "required": [
        "id",
        "userId",
        "groupId",
        "amount",
        "reason",
        "duration",
        "interestRate",
        "status",
        "createdAt"
      ],
      "properties": {
        "id": {
          "default": "(extensions.uuid_generate_v4())",
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "text",
          "type": "string"
        },
        "userId": {
          "description": "Note:\nThis is a Foreign Key to `users.id`.<fk table='users' column='id'/>",
          "format": "text",
          "type": "string"
        },
        "groupId": {
          "description": "Note:\nThis is a Foreign Key to `groups.id`.<fk table='groups' column='id'/>",
          "format": "text",
          "type": "string"
        },
        "amount": {
          "format": "double precision",
          "type": "number"
        },
        "reason": {
          "format": "text",
          "type": "string"
        },
        "duration": {
          "format": "integer",
          "type": "integer"
        },
        "interestRate": {
          "default": 10,
          "format": "double precision",
          "type": "number"
        },
        "status": {
          "default": "PENDING",
          "format": "text",
          "type": "string"
        },
        "createdAt": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.loans": {
      "name": "loans",
      "description": "loans",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/loans"
      }
    },
    "rowFilter.loans.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.userId": {
      "name": "userId",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.groupId": {
      "name": "groupId",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.amount": {
      "name": "amount",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.reason": {
      "name": "reason",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.duration": {
      "name": "duration",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.interestRate": {
      "name": "interestRate",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.status": {
      "name": "status",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.loans.createdAt": {
      "name": "createdAt",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}